+++ packages/varlet-cli/src/config/vite.config.ts
@@ -67,6 +67,7 @@ export function getBuildConfig(varletConfig: Record<string, any>): InlineConfig 
       outDir: SITE_OUTPUT_PATH,
       brotliSize: false,
       emptyOutDir: true,
+      cssTarget: 'chrome61',
       rollupOptions: {
         input: {
           main: resolve(SITE_DIR, 'index.html'),