+++ packages/varlet-ui/src/date-picker/DatePicker.vue
@@ -19,9 +19,7 @@
         ]"
         @click="clickEl('date')"
       >
-        <transition
-          :name="multiple ? '' : reverse ? 'var-date-picker-reverse-translatey' : 'var-date-picker-translatey'"
-        >
+        <transition :name="multiple ? '' : `var-date-picker${reverse ? '-reverse' : ''}-translatey`">
           <div :key="`${chooseYear}${chooseMonth?.index}`" v-if="type === 'month'">
             <slot name="range" :choose="getChoose.chooseRangeMonth" v-if="range">
               {{ getMonthTitle }}
@@ -47,15 +45,21 @@
         </transition>
       </div>
     </div>
-    <div class="var-date-picker-body">
+    <div
+      class="var-date-picker-body"
+      @touchstart="handleTouchstart"
+      @touchmove="handleTouchmove"
+      @touchend="handleTouchend"
+    >
       <transition name="var-date-picker-panel-fade">
         <year-picker-panel
           :component-props="componentProps"
           :preview="previewYear"
           @choose-year="getChooseYear"
-          v-if="isYearPanel"
+          v-if="getPanelType === 'year'"
         />
         <month-picker-panel
+          ref="monthPanelEl"
           :current="currentDate"
           :choose="getChoose"
           :preview="getPreview"
@@ -63,9 +67,10 @@
           :component-props="componentProps"
           @choose-month="getChooseMonth"
           @check-preview="checkPreview"
-          v-else-if="(!isYearPanel && type === 'month') || isMonthPanel"
+          v-else-if="getPanelType === 'month'"
         />
         <day-picker-panel
+          ref="dayPanelEl"
           :current="currentDate"
           :choose="getChoose"
           :preview="getPreview"
@@ -73,7 +78,7 @@
           :click-month="() => clickEl('month')"
           @choose-day="getChooseDay"
           @check-preview="checkPreview"
-          v-else-if="!isYearPanel && !isMonthPanel && type === 'date'"
+          v-else-if="getPanelType === 'date'"
         />
       </transition>
     </div>
@@ -88,9 +93,10 @@ import YearPickerPanel from './src/year-picker-panel.vue'
 import DayPickerPanel from './src/day-picker-panel.vue'
 import { props, MONTH_LIST, WEEK_HEADER } from './props'
 import { isArray, toNumber } from '../utils/shared'
+import { nextTickFrame } from '../utils/elements'
 import { pack } from '../locale'
-import type { Ref, ComputedRef, UnwrapRef } from 'vue'
-import type { MonthDict, Choose, Preview, WeekDict, ComponentProps } from './props'
+import type { Ref, ComputedRef, UnwrapRef, RendererNode } from 'vue'
+import type { MonthDict, Choose, Preview, WeekDict, ComponentProps, TouchDirection } from './props'
 
 export default defineComponent({
   name: 'VarDatePicker',
@@ -101,6 +107,10 @@ export default defineComponent({
   },
   props,
   setup(props) {
+    let startX = 0
+    let startY = 0
+    let checkType = ''
+    let touchDirection: TouchDirection | undefined
     const currentDate: string = dayjs().format('YYYY-MM-D')
     const [currentYear, currentMonth] = currentDate.split('-')
     const monthDes: MonthDict = MONTH_LIST.find((month) => month.index === currentMonth) as MonthDict
@@ -117,6 +127,8 @@ export default defineComponent({
     const chooseDays: Ref<Array<string>> = ref([])
     const chooseRangeMonth: Ref<Array<string>> = ref([])
     const chooseRangeDay: Ref<Array<string>> = ref([])
+    const monthPanelEl: Ref<RendererNode | null> = ref(null)
+    const dayPanelEl: Ref<RendererNode | null> = ref(null)
 
     const componentProps: UnwrapRef<ComponentProps> = reactive({
       allowedDates: props.allowedDates,
@@ -182,14 +194,29 @@ export default defineComponent({
       return `${weekName.slice(0, 3)}, ${monthName.slice(0, 3)} ${chooseDay.value}`
     })
 
+    const getPanelType: ComputedRef<string> = computed(() => {
+      if (isYearPanel.value) return 'year'
+
+      if (props.type === 'month' || isMonthPanel.value) return 'month'
+
+      if (props.type === 'date') return 'date'
+
+      return ''
+    })
+
+    const isUntouchable: ComputedRef<boolean> = computed(() => {
+      return !props.touchable || ['', 'year'].includes(getPanelType.value)
+    })
+
     const slotProps: ComputedRef<Record<string, string>> = computed(() => {
       const weekIndex = dayjs(`${chooseYear.value}-${chooseMonth.value?.index}-${chooseDay.value}`).day()
+      const date = chooseDay.value ? chooseDay.value?.padStart(2, '0') : ''
 
       return {
         week: `${weekIndex}`,
         year: chooseYear.value ?? '',
         month: chooseMonth.value?.index ?? '',
-        date: chooseDay.value ?? '',
+        date,
       }
     })
 
@@ -207,6 +234,39 @@ export default defineComponent({
         isYearPanel.value = false
         isMonthPanel.value = false
       }
+    }
+
+    const handleTouchstart = (event: TouchEvent) => {
+      if (isUntouchable.value) return
+
+      const { clientX, clientY } = event.touches[0]
+      startX = clientX
+      startY = clientY
+    }
+
+    const getDirection = (x: number, y: number): TouchDirection => {
+      return x >= y && x > 20 ? 'x' : 'y'
+    }
+
+    const handleTouchmove = (event: TouchEvent) => {
+      if (isUntouchable.value) return
+
+      const { clientX, clientY } = event.touches[0]
+      const x = clientX - startX
+      const y = clientY - startY
+
+      touchDirection = getDirection(Math.abs(x), Math.abs(y))
+      checkType = x > 0 ? 'prev' : 'next'
+    }
+
+    const handleTouchend = () => {
+      if (isUntouchable.value || touchDirection !== 'x') return
+
+      const componentRef = getPanelType.value === 'month' ? monthPanelEl : dayPanelEl
+      nextTickFrame(() => {
+        componentRef.value!.forwardRef(checkType)
+        resetState()
+      })
     }
 
     const updateRange = (date: string, type: string) => {
@@ -378,6 +438,13 @@ export default defineComponent({
       previewYear.value = yearValue
     }
 
+    const resetState = () => {
+      startY = 0
+      startX = 0
+      checkType = ''
+      touchDirection = undefined
+    }
+
     watch(
       () => props.modelValue,
       (value) => {
@@ -399,7 +466,11 @@ export default defineComponent({
       { immediate: true }
     )
 
+    watch(getPanelType, resetState)
+
     return {
+      monthPanelEl,
+      dayPanelEl,
       reverse,
       currentDate,
       chooseMonth,
@@ -410,12 +481,16 @@ export default defineComponent({
       isMonthPanel,
       getMonthTitle,
       getDateTitle,
+      getPanelType,
       getChoose,
       getPreview,
       componentProps,
       slotProps,
       formatRange,
       clickEl,
+      handleTouchstart,
+      handleTouchmove,
+      handleTouchend,
       getChooseDay,
       getChooseMonth,
       getChooseYear,