+++ packages/varlet-ui/src/pagination/example/ShowTotal.vue
@@ -39,6 +39,7 @@ export default {
 
 <style lang="less" scoped>
 ul {
+  margin: 0;
   padding: 10px 0;
   box-shadow: none;
 }