+++ packages/varlet-ui/src/pagination/example/SizeOption.vue
@@ -27,6 +27,7 @@ export default {
 
 <style lang="less" scoped>
 ul {
+  margin: 0;
   padding: 10px 0;
   box-shadow: none;
 }