+++ packages/varlet-ui/src/uploader/Uploader.vue
@@ -5,7 +5,7 @@
         class="var-uploader__file var-elevation--2"
         :class="[f.state === 'loading' ? 'var-uploader--loading' : null]"
         :key="f.id"
-        v-for="f in modelValue"
+        v-for="f in files"
         v-ripple="{ disabled: disabled || formDisabled || readonly || formReadonly || !ripple }"
         @click="preview(f)"
       >
@@ -135,6 +135,16 @@ export default defineComponent({
       // expose
       resetValidation,
     } = useValidation()
+
+    const files = computed(() => {
+      const { modelValue, hideList } = props
+
+      if (hideList) {
+        return []
+      }
+
+      return modelValue
+    })
 
     const preview = (varFile: VarFile) => {
       const { disabled, readonly, previewed } = props
@@ -315,6 +325,7 @@ export default defineComponent({
     )
 
     return {
+      files,
       showPreview,
       currentPreview,
       errorMessage,